+++ src/app/(app)/dashboard/business-profile/page.tsx
@@ -16,7 +16,6 @@ import DashboardCard from "../components/DashboardCard";
 import PageLoader from "@/app/(app)/components/PageLoader";
 import PageCard from "@/app/(app)/components/PageCard";
 import imageCompression from 'browser-image-compression';
-import FiveStarSpinner from "@/app/(app)/components/FiveStarSpinner";
 import { trackEvent, GA_EVENTS } from "@/utils/analytics";
 import { markTaskAsCompleted } from "@/utils/onboardingTasks";
 import WelcomePopup from "@/app/(app)/components/WelcomePopup";